+++ include/BDSDetectorConstruction.hh
@@ -172,6 +172,7 @@ private:
   BDSBeamline* placementBL; ///< Placement beam line.
   G4double     brho;        ///< Beam rigidity that accelerator will be constructed w.r.t.
   G4double     beta0;       ///< Beam relativistic beta that accelerator components use.
+  G4bool canSampleAngledFaces; ///< Whether the integrator set permits sampling elements with angled faces.
 };
 
 #endif